COL-2 Minor additions to the project with modifications to the current project, which only currently includes this task's code. Once the code for WorldState and Movement Systems are added, those will be the first commit to the task for that task.
This commit added a Menu() method to allow for Program.cs to be used as the primary test suite for the project.

As a result, the existing logic in Main(), which contained the Life Cycle Component tests, has been moved into a TestLifecycle() method and can now be  run by selecting the first option in the menu.

To quit the test suite, select the 'Q' option.

This test suite should be moved into its own test class at some point in the future, however, to keep Program.cs as mostly a driver program where several options are available. This way, the Test Suite can still be used and updated for individual component tests or testing certain systems while still allowing the full project to be run. This can be discussed for a later Sprint when systems begin to use multiple components and begin to interact more often.

NOTE: Use Thread.Sleep(500) on any options added after the output confirming the start of tests and before the actual test method call to keep tests from immediately starting to spit out output. This allows for better tracking of output from the beginning of the tests with potentially large amounts of output.